+++ include/fc/io/raw.hpp
@@ -180,9 +180,9 @@ namespace fc {
 
     // std::vector<char>
     template<typename Stream> inline void pack( Stream& s, const std::vector<char>& value ) { 
-      pack( s, unsigned_int(value.size()) );
+      pack( s, unsigned_int((uint32_t)value.size()) );
       if( value.size() )
-        s.write( &value.front(), value.size() );
+        s.write( &value.front(), (uint32_t)value.size() );
     }
     template<typename Stream> inline void unpack( Stream& s, std::vector<char>& value ) { 
       unsigned_int size; unpack( s, size );
@@ -194,7 +194,7 @@ namespace fc {
 
     // fc::string
     template<typename Stream> inline void pack( Stream& s, const fc::string& v )  {
-      pack( s, unsigned_int(v.size()) );     
+      pack( s, unsigned_int((uint32_t)v.size()));     
       if( v.size() ) s.write( v.c_str(), v.size() );
     }
 
@@ -312,7 +312,7 @@ namespace fc {
 
     template<typename Stream, typename T>
     inline void pack( Stream& s, const std::unordered_set<T>& value ) {
-      pack( s, unsigned_int(value.size()) );
+      pack( s, unsigned_int((uint32_t)value.size()) );
       auto itr = value.begin();
       auto end = value.end();
       while( itr != end ) {
@@ -349,7 +349,7 @@ namespace fc {
 
     template<typename Stream, typename K, typename V>
     inline void pack( Stream& s, const std::unordered_map<K,V>& value ) {
-      pack( s, unsigned_int(value.size()) );
+      pack( s, unsigned_int((uint32_t)value.size()) );
       auto itr = value.begin();
       auto end = value.end();
       while( itr != end ) {
@@ -373,7 +373,7 @@ namespace fc {
     }
     template<typename Stream, typename K, typename V>
     inline void pack( Stream& s, const std::map<K,V>& value ) {
-      pack( s, unsigned_int(value.size()) );
+      pack( s, unsigned_int((uint32_t)value.size()) );
       auto itr = value.begin();
       auto end = value.end();
       while( itr != end ) {
@@ -398,7 +398,7 @@ namespace fc {
 
     template<typename Stream, typename T>
     inline void pack( Stream& s, const std::vector<T>& value ) {
-      pack( s, unsigned_int(value.size()) );
+      pack( s, unsigned_int((uint32_t)value.size()) );
       auto itr = value.begin();
       auto end = value.end();
       while( itr != end ) {
@@ -422,7 +422,7 @@ namespace fc {
 
     template<typename Stream, typename T>
     inline void pack( Stream& s, const std::set<T>& value ) {
-      pack( s, unsigned_int(value.size()) );
+      pack( s, unsigned_int((uint32_t)value.size()) );
       auto itr = value.begin();
       auto end = value.end();
       while( itr != end ) {
